refactor: extract homepage sections into dedicated Svelte components, polish ticker animation and mobile nav
