+++ src/app.css
@@ -138,7 +138,7 @@ body {
 
 @keyframes ticker-scroll {
 	0% {
-		transform: translateX(0);
+		transform: translateX(100vw);
 	}
 	100% {
 		transform: translateX(-50%);
@@ -146,7 +146,9 @@ body {
 }
 
 .animate-ticker {
-	animation: ticker-scroll 80s linear infinite;
+	animation: ticker-scroll 90s linear infinite;
+	animation-delay: 0.3s;
+	animation-fill-mode: both;
 }
 
 .animate-ticker:hover {
