+++ src/routes/components/ContactSection.svelte
@@ -1,0 +1,28 @@
+<section id="contact" class="border-b border-(--c-line-10) px-6 py-10 md:px-12 md:py-12 lg:px-24">
+	<p class="mb-8 font-mono text-[10px] uppercase tracking-widest text-(--c-accent)">
+		// STATUS: ACCEPTING WORK Q3—Q4 {new Date().getFullYear()}
+	</p>
+
+	<h2
+		class="mb-12 font-display text-[clamp(3rem,10vw,8rem)] font-bold uppercase leading-none tracking-tight text-(--c-text)"
+	>
+		Got something<br /><em class="not-italic text-(--c-accent)">worth</em><br />building?
+	</h2>
+
+	<div class="flex flex-col sm:flex-row gap-3">
+		<a
+			href="mailto:aleksebastian@outlook.com"
+			class="inline-flex items-center gap-3 border-2 border-(--c-line-20) px-6 py-4 font-mono text-xs uppercase tracking-widest text-(--c-text) transition-all duration-200 group hover:border-(--c-accent) hover:bg-(--c-accent) hover:text-(--c-text-inverse)"
+		>
+			ALEKSEBASTIAN@OUTLOOK.COM
+			<span class="group-hover:translate-x-1 transition-transform">→</span>
+		</a>
+		<a
+			href="/Alek%20Ortiz%20resume.pdf"
+			download="Alek Ortiz resume.pdf"
+			class="inline-flex items-center gap-3 border-2 border-(--c-line-10) px-6 py-4 font-mono text-xs uppercase tracking-widest text-(--c-text-40) transition-all duration-200 hover:border-(--c-line-30) hover:text-(--c-text-70)"
+		>
+			DOWNLOAD CV [.PDF]
+		</a>
+	</div>
+</section>
