+++ src/routes/components/SiteFooter.svelte
@@ -1,0 +1,15 @@
+<footer
+	class="px-6 md:px-12 lg:px-24 py-6 flex flex-col sm:flex-row justify-between items-start sm:items-center gap-3"
+>
+	<span class="font-mono text-[10px] uppercase tracking-widest text-(--c-text-25)">
+		© {new Date().getFullYear()} ALEK ORTÍZ
+	</span>
+	<div class="flex flex-col sm:flex-row gap-3 sm:gap-8">
+		<span class="font-mono text-[10px] uppercase tracking-widest text-(--c-text-25)">
+			BUILT WITH SVELTE + 🎧
+		</span>
+		<span class="font-mono text-[10px] uppercase tracking-widest text-(--c-text-25)">
+			v2.0 / {new Date().getFullYear()}.04
+		</span>
+	</div>
+</footer>
